Some refatoring of code. Improving of visuals in MainWindow.xml. Bug Fixes: 	- Old Image wouldn't be removed when the same addon would be installed 	twice.

+++ AddonMaster.GUI/MainWindow.xaml.cs
@@ -1,57 +1,55 @@
 ﻿using AddonMaster.Core.Data;
-using AddonMaster.Core.Data.Entities;
 using AddonMaster.GUI.ViewModels;
-using MahApps.Metro.Controls;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.IO;
 using System.Linq;
-using System.Threading;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
 
 namespace AddonMaster.GUI
 {
-    public partial class MainWindow : MetroWindow
+    public partial class MainWindow
     {
-        private DatabaseManager dbManager;
+        private readonly DatabaseManager _dbManager;
 
         public MainWindow(string addonFolderPath)
         {
             InitializeComponent();
 
-            dbManager = new DatabaseManager(addonFolderPath);
+            _dbManager = new DatabaseManager(addonFolderPath);
 
             UpdateListBoxOnMainWindow();
         }
 
         private void imgAdd_MouseDown(object sender, MouseButtonEventArgs e)
         {
-            var addonWindow = new AddAddonWindow(this, dbManager);
-            addonWindow.Owner = this;
+            var addonWindow = new AddAddonWindow(this, _dbManager) {Owner = this};
             addonWindow.Show();
         }
 
         private void imgUpdate_MouseDown(object sender, MouseButtonEventArgs e)
         {
-            var addonViewModel = (sender as Image).DataContext as AddonViewModel;
+            var addonViewModel = (sender as Image)?.DataContext as AddonViewModel;
 
-            var worker = new BackgroundWorker();
-            worker.WorkerReportsProgress = true;
+            var worker = new BackgroundWorker {WorkerReportsProgress = true};
 
-            worker.DoWork += (object x, DoWorkEventArgs y) =>
+            worker.DoWork += (x, y) =>
             {
-                (x as BackgroundWorker).ReportProgress(0);
-                dbManager.RemoveAddon(addonViewModel.Addon.ID);
-                dbManager.AddAddon(addonViewModel.Addon.DownloadUrl, worker);
+                (x as BackgroundWorker)?.ReportProgress(0);
+                if (addonViewModel != null)
+                {
+                    _dbManager.RemoveAddon(addonViewModel.Addon.Id);
+                    _dbManager.AddAddon(addonViewModel.Addon.DownloadUrl, worker);
+                }
             };
 
-            worker.ProgressChanged += (object x, ProgressChangedEventArgs y) => Spinner.Visibility = Visibility.Visible;
+            worker.ProgressChanged += (x, y) => Spinner.Visibility = Visibility.Visible;
 
-            worker.RunWorkerCompleted += (object x, RunWorkerCompletedEventArgs y) =>
+            worker.RunWorkerCompleted += (o, args) =>
             {
-                File.Delete(addonViewModel.Addon.ImagePath);
+                if (addonViewModel != null) File.Delete(addonViewModel.Addon.ImagePath);
                 UpdateListBoxOnMainWindow();
                 Spinner.Visibility = Visibility.Collapsed;
             };
@@ -61,14 +59,19 @@ namespace AddonMaster.GUI
 
         private void imgRemove_MouseDown(object sender, MouseButtonEventArgs e)
         {
-            var addonViewModel = (sender as Image).DataContext as AddonViewModel;
-            var oldImagePath = string.Empty;
+            var addonViewModel = (sender as Image)?.DataContext as AddonViewModel;
 
-            var worker = new BackgroundWorker();
-            worker.WorkerReportsProgress = true;
+            var worker = new BackgroundWorker {WorkerReportsProgress = true};
 
-            worker.DoWork += (object x, DoWorkEventArgs y) => dbManager.RemoveAddon(addonViewModel.Addon.ID);
-            worker.RunWorkerCompleted += (object x, RunWorkerCompletedEventArgs y) => { File.Delete(addonViewModel.Addon.ImagePath); UpdateListBoxOnMainWindow(); };
+            worker.DoWork += (x, y) =>
+            {
+                if (addonViewModel != null) _dbManager.RemoveAddon(addonViewModel.Addon.Id);
+            };
+            worker.RunWorkerCompleted += (x, y) =>
+            {
+                if (addonViewModel != null) File.Delete(addonViewModel.Addon.ImagePath);
+                UpdateListBoxOnMainWindow();
+            };
 
             worker.RunWorkerAsync();
         }
@@ -77,69 +80,54 @@ namespace AddonMaster.GUI
         {
             var addonViews = new List<AddonViewModel>();
 
-            dbManager.GetAddons().OrderBy(x => x.Name).ToList().ForEach(x =>
+            _dbManager.GetAddons().OrderBy(x => x.Name).ToList().ForEach(x =>
             {
                 addonViews.Add(new AddonViewModel { Addon = x });
             });
 
-            lbAddonList.ItemsSource = addonViews;
-
-            lbAddonList.Items.Refresh();
-
-            if (lbAddonList.Items.Count != 0)
-            {
-                lblEmptyList.Visibility = Visibility.Collapsed;
-            }
-            else
-            {
-                lblEmptyList.Visibility = Visibility.Visible;
-            }
+            LblAddonList.ItemsSource = addonViews;
+            LblAddonList.Items.Refresh();
+            LblEmptyList.Visibility = LblAddonList.Items.Count != 0 ? Visibility.Collapsed : Visibility.Visible;
         }
 
         private void imgReport_MouseDown(object sender, MouseButtonEventArgs e)
         {
-            var reportWindow = new ReportWindow();
-            reportWindow.Owner = this;
-
+            var reportWindow = new ReportWindow {Owner = this};
             reportWindow.Show();
         }
 
         #region UI Stuff
         private void Window_MouseDown(object sender, MouseButtonEventArgs e)
         {
-            if (e.ChangedButton == MouseButton.Left)
+            if (e.ChangedButton != MouseButton.Left) return;
+            try
             {
-                try
-                {
-                    DragMove();
-                }
-                catch { }
+                DragMove();
+            }
+            catch
+            {
+                // ignored
             }
         }
         #endregion
 
         private void ImageAwesome_MouseDown(object sender, MouseButtonEventArgs e)
         {
-            if ((sender as Image).ToolTip as string == "Close")
+            if ((sender as Image)?.ToolTip as string == "Close")
             {
                 Close();
             }
-            else if ((sender as Image).ToolTip as string == "Minimize")
+            else if ((sender as Image)?.ToolTip as string == "Minimize")
             {
                 WindowState = WindowState.Minimized;
             }
-            else if ((sender as Image).ToolTip as string == "Help")
+            else if ((sender as Image)?.ToolTip as string == "Help")
             {
                 new InfoWindow(this)
                 {
                     Owner = this
                 }.Show();
             }
-        }
-
-        private void Grid_MouseEnter(object sender, MouseEventArgs e)
-        {
-
         }
     }
 }
